+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+- Add debug options for sfdx-hardis commands
+- Add new sandbox refresh commands
+- New UI to update VsCode SFDX-Hardis configuration
+- Fix issue on MacOS with background commands
+
+## [6.0.9 (pre-release)] 2024-08-14
+
+- Add a new mode (by default) to execute sfdx-hardis commands in background, not in terminal
+
 ## [6.0.8 (pre-release)] 2024-08-12
 
 - Command execution: Add simple/advanced mode